Add new endpoints and update DTOs and services
- Changed `OrgSerial` in `ReservationsDto` from `int?` to `double?`.
- Added `GetDisplayDetails` endpoint to `DisplayController`.
- Added `GetServingReservationsBasedOnOrgId` endpoint to `ReservationsController`.
- Updated `IDisplayServices` and `IReservationServices` interfaces with new methods.
- Updated `DisplayProfile`, `DisplayAdvertsProfile`, and `DisplayCounterProfile` with new mappings and simplified type names.
- Added `GetDisplayDetails` method to `DisplayServices`.
- Updated `ReservationService`:
  - Changed statuses from `Pending` to `Waiting` and `Completed` to `Served`.
  - Handled exceptions for SMS sending.
  - Added `GetServingReservationsBasedOnOrgId` method.
  - Updated `CallNextInQueue` and reservation statistics to use new statuses.
- Added `DisplayAdvertsTinyDto`, `DisplayCountersTinyDto`, and `DisplayDetailsDto` classes.

+++ ApiContracts/Reservation/ReservationsDto.cs
@@ -39,7 +39,7 @@ public class ReservationsDto
     [JsonPropertyName("reservationType")]
     public string? ReservationType { get; set; }
     [JsonPropertyName("orgSerial")]
-    public int? OrgSerial { get; set; }
+    public double? OrgSerial { get; set; }
     [JsonPropertyName("transferedAt")]
     public DateTime? TransferedAt { get; set; }
     [JsonPropertyName("orgId")]
@@ -48,5 +48,6 @@ public class ReservationsDto
     public string OrgName { get; set; } = null!;
     [JsonPropertyName("services")]
     public List< string>? Services { get; set; }
+
 
 }
